Move fixtures and helper functions to repobee_plug
