+++ src/repobee_plug/testhelpers/resources/templates/task-3/.gitignore
