+++ tests/new_integration_tests/conftest.py
@@ -1,0 +1,6 @@
+# flake8: noqa
+from repobee_plug.testhelpers.fixtures import (
+    platform_dir,
+    platform_url,
+    with_student_repos,
+)
