+++ tests/new_integration_tests/test_issues.py
@@ -1,0 +1,1 @@
+"""Tests for the issues category of commands."""
